Facotred out search payload type
